Should be vs is adjacent data API cleanup
In array object model distinguish between APIs:
- should be adjacent
- is adjacent

The former is used early when the object is being constructed, and
dataAddr is yet to be initialized. API makes a decision based on size in
bytes explicitely provided or read from the array heder, and is used to
determine if data should be allocated in main heap or offheap.

The latter is used at later points in time (like GC, JNI Critical API)
when object is fully initialized. Then the check is about actual
dataAddr value (if it points to the end of the header or not). This is
faster and simpler check and is prefered, unless dataAddr is not
initialized yet.

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>